La reprise de la correction des erreurs dans la gestion des employes

+++ page/employe/employes.php
@@ -6,7 +6,8 @@ $modifierApres =0;
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     // Récupérez les données envoyées via POST
    
-    if($_POST['action'] === 'Sauvegarder'){
+    if($_POST['action'] === 'Sauvegarder')
+    {
         //Recupération des données
         $idEmploye = intval($_POST['id_emp']);
         $nomEmploye = $_POST['nom_emp'];
@@ -25,47 +26,44 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         //Téléchargement de la photo
         $dossierTelechargement = "C:/wamp64/www/DEV_VERSION_2/image/photo_employe/";
         $cheminFichierCible ="";
-        if(isset($_FILES)){
+
+        if (isset($_FILES)) {
             $fichierTelecharge = $_FILES['photo_emp'];
             $nomFichier = basename($fichierTelecharge['name']);
             $cheminFichierCible = $dossierTelechargement . $nomFichier;
-
+        
             // Vérifiez si le dossier de destination existe, sinon créez-le
             if (!file_exists($dossierTelechargement)) {
                 mkdir($dossierTelechargement, 0777, true);
             }
-            // Déplacer le fichier téléchargé vers le dossier cible
-            if (!move_uploaded_file($fichierTelecharge['tmp_name'], $cheminFichierCible)) {
-                $nomFichier = $_POST['photo_emp_ent'];
-                $cheminFichierCible = $dossierTelechargement.$nomFichier;
-                
-            } 
-        }
         
-       
-        $idService =getOneServiceWithName($serviceEmploye)[0]["id_serv"];
-    
-        if ($modifierAvant === 0 && $modifierApres === 0){
-            $answer = updateEmploye($idEmploye, $nomEmploye, $prenomEmploye, $sexeEmploye, $nationaliteEmploye,$lieuResEmploye, $salaireEmploye
-                                    , $emailEmploye, $contactEmploye, $dateNaisEmploye, 
-                            $dateEmbauEmploye, $niveauEtudeEmploye,$idService ,$cheminFichierCible);
-            echo $answer;
-
-            if($answer){
-                echo "supper";
-                $modifierApres =1;
-            }else{
-                $modifierApres = 0;
+            // Vérifiez si le fichier existe déjà dans le dossier cible
+            if (!file_exists($cheminFichierCible)) {
+                // Déplacez le fichier téléchargé vers le dossier cible
+                if (!move_uploaded_file($fichierTelecharge['tmp_name'], $cheminFichierCible)) {
+                    // Si le déplacement échoue, utiliser le nom de la photo existante passée via le formulaire
+                    $nomFichier = $_POST['photo_emp_ent'];
+                    $cheminFichierCible = $dossierTelechargement . $nomFichier;
+                }
+            } else {
+                // Si le fichier existe déjà, utilisez le nom de la photo existante passée via le formulaire
+                $nomFichier = $_POST['photo_emp_ent'];
+                $cheminFichierCible = $dossierTelechargement . $nomFichier;
             }
         }
-        
-        
-        
 
+        $idService =getOneServiceWithName($serviceEmploye)[0]["id_serv"];
         
-    }elseif($_POST['action'] === "enregistrernouvel"){
         
-        $idEmploye = $_POST['id_emp_new'];
+        $answer = updateEmploye($idEmploye, $nomEmploye, $prenomEmploye, $sexeEmploye, $nationaliteEmploye,$lieuResEmploye, $salaireEmploye
+                                    , $emailEmploye, $contactEmploye, $dateNaisEmploye, 
+                            $dateEmbauEmploye, $niveauEtudeEmploye,$idService ,$cheminFichierCible);
+       
+            
+    }
+    elseif($_POST['action'] === "enregistrernouvel")
+    {   
+  
         $nomEmploye = $_POST['nom_emp_new'];
         $prenomEmploye = $_POST['prenom_emp_new'];
         $sexeEmploye = $_POST['Sexe_new'];
@@ -76,13 +74,13 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         $contactEmploye = $_POST['contact_emp_new'];
         $dateNaisEmploye = $_POST['date_nais_emp_new'];
         $dateEmbauEmploye = $_POST['date_embau_emp_new'];
-        $serviceEmploye = $_POST['nom_serv_new'];
+        $serviceEmploye = $_POST['id_serv_new'];
         $niveauEtudeEmploye = $_POST['niveau_etu_emp_new'];
 
         //Téléchargement de la photo
         $dossierTelechargement = 'C:/wamp64/www/DEV_VERSION_2/image/photo_employe/';
         $cheminFichierCible ="";
-        if(isset($_FILES)){
+        if(isset($_FILES) && $_FILES['photo_emp_new']['name']!==''){
             $fichierTelecharge = $_FILES['photo_emp_new'];
             $nomFichier = basename($fichierTelecharge['name']);
             $cheminFichierCible = $dossierTelechargement . $nomFichier;
@@ -91,22 +89,38 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             if (!file_exists($dossierTelechargement)) {
                 mkdir($dossierTelechargement, 0777, true);
             }
-            // Déplacer le fichier téléchargé vers le dossier cible
-            if (!move_uploaded_file($fichierTelecharge['tmp_name'], $cheminFichierCible)) {
-                $cheminFichierCible = $dossierTelechargement."photo_de_profile.jpeg";
-                
+            
+            // Vérifier si le fichier existe déjà dans le répertoire de destination
+            if (!file_exists($cheminFichierCible)) {
+                // Déplacer le fichier téléchargé vers le dossier cible si le fichier n'y ait pas
+                //Sinon on prend la photo par défaut
+                if (!move_uploaded_file($fichierTelecharge['tmp_name'], $cheminFichierCible)) {
+                    $cheminFichierCible = "";
+                }
             }
 
         }
-        echo $cheminFichierCible;
-        var_dump($_FILES);
-        var_dump($_POST);
+        else 
+        {
+
+            $cheminFichierCible = $dossierTelechargement."photo_de_profile.jpeg";
+            
+        }
+        
+  
+       insertEmploye($nomEmploye, $prenomEmploye, $sexeEmploye, $nationaliteEmploye,
+        $lieuResEmploye, $salaireEmploye, $emailEmploye, $contactEmploye, $dateNaisEmploye, $dateEmbauEmploye, 
+        $niveauEtudeEmploye, $serviceEmploye, $cheminFichierCible);
+  
     }
     elseif($_POST['action']==="supprimerEmployeConfirme"){
         var_dump($_POST);
     }
     elseif($_POST['action']==="exporterEmployeInfo"){
         var_dump($_POST);
+    }
+    else{
+        echo "NNNNNNNNNNNNNNNNNNNNNNNNNNNN33";
     }
